+++ app/src/main/java/woowacourse/movie/presentation/purchaseConfirmation/PurchaseConfirmationActivity.kt
@@ -69,29 +69,3 @@ class PurchaseConfirmationActivity :
     }
 }
 
-interface PurchaseConfirmationContract {
-    interface View {
-        fun showReservation(reservation: Reservation)
-
-        fun showError()
-    }
-
-    interface Presenter {
-        fun loadReservation(reservationId: Long)
-    }
-}
-
-class PurchaseConfirmationPresenter(
-    private val repository: MovieRepository,
-    private val view: PurchaseConfirmationContract.View
-) : PurchaseConfirmationContract.Presenter {
-    override fun loadReservation(reservationId: Long) {
-        thread {
-            repository.loadReservedMovie(reservationId).onSuccess {
-                view.showReservation(it)
-            }.onFailure {
-                view.showError()
-            }
-        }.join()
-    }
-}
